requirements.txt
flask>=3.0.0
smartapi-python>=1.3.6
pyotp>=2.9.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
gunicorn>=21.0.0
